+++ First/gotoeat/gotoeat-api/src/main/java/kr/co/first/gotoeat/domain/RestaurantNotFoundException.java
@@ -3,7 +3,7 @@ package kr.co.first.gotoeat.domain;
 public class RestaurantNotFoundException extends RuntimeException
 {
 
-    public RestaurantNotFoundException(int id)
+    public RestaurantNotFoundException(long id)
     {
         super("Could Not Found Restaurant" + id);
     }
